handle problematic imports
